Unify export pipeline into single FlavorExporter strategy
Replace 6 separate exporter strategies (SchemeExporter, RegistryExporter,
NistExporter, IeeeExporter, DataClassExporter, ItuExporter) with a single
unified FlavorExporter that uses the self-describing type interface on each
flavor module (identifier_types, all_typed_stages, locate_type, locate_stage).

- Delete 6 strategy subclasses and their specs
- Simplify Exporter to a single FLAVORS list + FlavorExporter
- Add fallback namespace scanning for flavors without type declarations
- Add identifier_types to API module (was missing)
- Export produces data for all 23 flavors with 170+ identifier types
