+++ lib/pubid/export.rb
@@ -6,12 +6,6 @@ module Pubid
     autoload :TypedStageResult, "#{__dir__}/export/result"
     autoload :FlavorResult, "#{__dir__}/export/result"
     autoload :FlavorExporter, "#{__dir__}/export/flavor_exporter"
-    autoload :SchemeExporter, "#{__dir__}/export/scheme_exporter"
-    autoload :IeeeExporter, "#{__dir__}/export/ieee_exporter"
-    autoload :RegistryExporter, "#{__dir__}/export/registry_exporter"
-    autoload :DataClassExporter, "#{__dir__}/export/data_class_exporter"
-    autoload :NistExporter, "#{__dir__}/export/nist_exporter"
-    autoload :ItuExporter, "#{__dir__}/export/itu_exporter"
     autoload :Auditor, "#{__dir__}/export/auditor"
     autoload :Exporter, "#{__dir__}/export/exporter"
   end
