+++ lib/pubid/export/exporter.rb
@@ -2,53 +2,15 @@
 
 module Pubid
   module Export
-    # Top-level orchestrator. Iterates flavors, selects the appropriate
-    # extraction strategy, and produces the combined result.
+    # Top-level orchestrator. Iterates flavors and delegates to FlavorExporter.
     #
-    # Open/Closed: Adding a new flavor only requires registering it in
-    # FLAVOR_STRATEGIES; no existing code changes.
+    # Open/Closed: Adding a new flavor only requires adding it to FLAVORS;
+    # no existing code changes.
     class Exporter
       FLAVORS = %i[
         iso iec ieee nist bsi itu cen_cenelec etsi ansi astm ashrae asme
         ccsds cie csa jis jcgm oiml idf api amca plateau sae
       ].freeze
-
-      # Maps each flavor to its extraction strategy class.
-      # When a flavor follows a new pattern, add a new strategy class here.
-      FLAVOR_STRATEGIES = {
-        iso: :scheme,
-        iec: :scheme,
-        ieee: :ieee,
-        nist: :nist,
-        bsi: :registry,
-        itu: :itu,
-        cen_cenelec: :registry,
-        etsi: :data_class,
-        ansi: :scheme,
-        astm: :scheme,
-        ashrae: :scheme,
-        asme: :scheme,
-        ccsds: :scheme,
-        cie: :scheme,
-        csa: :scheme,
-        jis: :scheme,
-        jcgm: :scheme,
-        oiml: :scheme,
-        idf: :scheme,
-        api: :scheme,
-        amca: :scheme,
-        plateau: :data_class,
-        sae: :scheme,
-      }.freeze
-
-      STRATEGY_CLASSES = {
-        scheme: SchemeExporter,
-        registry: RegistryExporter,
-        data_class: DataClassExporter,
-        nist: NistExporter,
-        ieee: IeeeExporter,
-        itu: ItuExporter,
-      }.freeze
 
       # Export all flavors to a hash suitable for JSON serialization.
       # @return [Hash{String => Hash}]
@@ -56,11 +18,7 @@ module Pubid
         require "pubid"
 
         FLAVORS.each_with_object({}) do |flavor, result|
-          strategy_name = FLAVOR_STRATEGIES[flavor]
-          strategy_class = STRATEGY_CLASSES[strategy_name]
-          next unless strategy_class
-
-          exporter = strategy_class.new(flavor)
+          exporter = FlavorExporter.new(flavor)
           flavor_result = exporter.export
           next unless flavor_result
